oppdaterer snapshots og legger til noen tester for komponenter som ikke lenger er dekket etter at valgt tråd og valgt sakstema bare holdes i url
